precalc halfsize for screenscaling

+++ GameSoftware3D_Data.h
@@ -138,7 +138,7 @@ namespace game
 		uint32_t* colorBuffer = nullptr;
 		float_t* depthBuffer = nullptr;
 		Pointi size;
-		Pointf halfSize;
+		Pointi halfSize;
 		uint32_t totalBufferSize = 0;
 	};
